relocate Tensor_impl
